requirements.txt
pandas
SQLAlchemy
psycopg2
python-dotenv